New continuous minimum ETA calculation
This more closely implements the idea that intense cognitive work is
more productive than average at the start of the day and that
productivity falls off a cliff after roughly 4 hours.  This accurately
predicts that spreading work over more days, but shorter, yields more
value.

The specific ratios chosen in output()'s block_outputs[] are so that 4.5
input hours yield 7.5 hours of output (less aggressive than the 4->8
rule of thumb) and that "overtime" nullifies this around the 9-hour
mark.